# api/openapi.yaml
openapi: 3.1.0
info:
  title: ChatP2P Backend API
  version: 0.1.0
  description: Backend-only private chat service. Frontend teams should treat this file as the REST API contract.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/docs/openapi.yaml:
    get:
      tags:
        - Docs
      summary: Get latest OpenAPI document
      operationId: getOpenAPIDocument
      description: Returns the current `api/openapi.yaml` file from the running server's repository checkout.
      responses:
        "200":
          description: Latest OpenAPI YAML document.
          content:
            application/yaml:
              schema:
                type: string
            text/yaml:
              schema:
                type: string
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/docs/recent:
    get:
      tags:
        - Docs
      summary: Get recently updated documentation
      operationId: listRecentDocs
      description: Returns recently modified Markdown files from `docs/`, including full content for quick frontend display.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Maximum number of recent docs to return.
      responses:
        "200":
          description: Recently updated documentation files.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentDocsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/ws:
    get:
      tags:
        - Realtime
      summary: Open WebSocket event stream
      operationId: openRealtimeWebSocket
      description: "Browser clients may pass `accessToken` as a query parameter. Non-browser clients should prefer `Authorization: Bearer <token>`."
      parameters:
        - name: accessToken
          in: query
          required: false
          schema:
            type: string
          description: Bearer token for clients that cannot set custom WebSocket headers.
      responses:
        "101":
          description: WebSocket upgrade accepted. Events use `RealtimeEvent` envelopes.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/sync:
    get:
      tags:
        - Sync
      summary: Sync state after reconnect
      operationId: syncState
      description: Returns the current conversation list and a page of messages created or updated after the supplied sync cursor. Prefer `cursor` for stable resume and pagination.
      security:
        - bearerAuth: []
      parameters:
        - name: since
          in: query
          required: false
          deprecated: true
          schema:
            type: string
            format: date-time
          description: Legacy sync timestamp. Prefer `cursor` for stable resume. Do not combine with `cursor`.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque sync cursor returned by the previous response's `nextCursor`. Do not combine with `since`.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of conversations and changed messages returned on this page.
      responses:
        "200":
          description: Synchronization snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/register:
    post:
      tags:
        - Auth
      summary: Register a user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User was created and authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Login with username and password
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Credentials were accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/sticker-packs/{packId}/stickers:
    get:
      tags:
        - Stickers
      summary: List stickers in a pack
      operationId: listStickers
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: string
          example: classic-faces
      responses:
        "200":
          description: Stickers in the requested pack.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/stickers/{stickerId}/asset.svg:
    get:
      tags:
        - Stickers
      summary: Get sticker SVG asset
      operationId: getStickerAsset
      parameters:
        - name: stickerId
          in: path
          required: true
          schema:
            type: string
          example: classic-smile
      responses:
        "200":
          description: Sticker SVG image.
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users:
    get:
      tags:
        - Users
      summary: Search users
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 32
          description: Matches username or display name.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Matched users, excluding the current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current authenticated user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags:
        - Users
      summary: Update current user profile
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: Updated authenticated user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/friend-requests:
    post:
      tags:
        - Friends
      summary: Send a friend request
      operationId: sendFriendRequest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFriendRequestRequest"
      responses:
        "201":
          description: Friend request was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      tags:
        - Friends
      summary: List friend requests
      operationId: listFriendRequests
      security:
        - bearerAuth: []
      parameters:
        - name: box
          in: query
          required: false
          schema:
            type: string
            enum:
              - incoming
              - outgoing
              - all
            default: incoming
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - pending
              - accepted
              - declined
            default: pending
      responses:
        "200":
          description: Friend requests for the selected box and status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequestListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/friend-requests/{id}/accept:
    post:
      tags:
        - Friends
      summary: Accept a friend request
      operationId: acceptFriendRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FriendRequestId"
      responses:
        "200":
          description: Friend request was accepted and a direct conversation is available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequestActionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/friend-requests/{id}/decline:
    post:
      tags:
        - Friends
      summary: Decline a friend request
      operationId: declineFriendRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FriendRequestId"
      responses:
        "200":
          description: Friend request was declined.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/friends:
    get:
      tags:
        - Friends
      summary: List current user's friends
      operationId: listFriends
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user's friends.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/friends/{userId}:
    delete:
      tags:
        - Friends
      summary: Remove a friend
      operationId: removeFriend
      description: Removes the accepted friendship between the current user and the target user. Existing conversations and messages are not deleted.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Friendship was removed.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/blocks:
    get:
      tags:
        - Friends
      summary: List blocked users
      operationId: listBlockedUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users blocked by the current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedUserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/blocks/{userId}:
    post:
      tags:
        - Friends
      summary: Block a user
      operationId: blockUser
      description: Blocks the target user, removes any friendship and friend-request state between the two users, and prevents new friend requests or direct conversation entry. Existing conversations and messages are not deleted.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "201":
          description: User was blocked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags:
        - Friends
      summary: Unblock a user
      operationId: unblockUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: User was unblocked.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations:
    get:
      tags:
        - Conversations
      summary: List current user's conversations
      operationId: listConversations
      security:
        - bearerAuth: []
      parameters:
        - name: before
          in: query
          required: false
          deprecated: true
          schema:
            type: string
            format: date-time
          description: Legacy timestamp filter for conversation lists. Prefer `cursor` for stable pagination.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque pagination cursor returned by the previous list response. Do not combine with `before`.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: includeArchived
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Include conversations archived by the current user. Archived conversations are hidden by default.
      responses:
        "200":
          description: Conversations ordered by pinned state, pin time, then latest activity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/conversations/direct:
    post:
      tags:
        - Conversations
      summary: Get or create a direct conversation with a friend
      operationId: createDirectConversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DirectConversationRequest"
      responses:
        "200":
          description: Existing or newly created direct conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations/group:
    post:
      tags:
        - Conversations
      summary: Create a group conversation
      operationId: createGroupConversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupConversationRequest"
      responses:
        "201":
          description: Group conversation was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations/{conversationId}:
    patch:
      tags:
        - Conversations
      summary: Rename a group conversation
      operationId: renameGroupConversation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameGroupConversationRequest"
      responses:
        "200":
          description: Group conversation was renamed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations/{conversationId}/settings:
    patch:
      tags:
        - Conversations
      summary: Update current user's conversation settings
      operationId: updateConversationSettings
      description: Updates per-user settings for an existing conversation membership. Send an empty `mutedUntil` string to clear mute.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationSettingsUpdateRequest"
      responses:
        "200":
          description: Updated conversation settings for the current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationSettings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations/{conversationId}/owner:
    patch:
      tags:
        - Conversations
      summary: Transfer group ownership
      operationId: transferGroupConversationOwner
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferGroupConversationOwnerRequest"
      responses:
        "200":
          description: Group owner was transferred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations/{conversationId}/members:
    post:
      tags:
        - Conversations
      summary: Add members to a group conversation
      operationId: addGroupConversationMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGroupConversationMembersRequest"
      responses:
        "200":
          description: Members were added to the group conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/conversations/{conversationId}/members/{userId}:
    delete:
      tags:
        - Conversations
      summary: Remove a member from a group conversation
      operationId: removeGroupConversationMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Member was removed from the group conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations/{conversationId}/leave:
    post:
      tags:
        - Conversations
      summary: Leave a group conversation
      operationId: leaveGroupConversation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Current user left the group conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversations/{conversationId}/messages:
    post:
      tags:
        - Conversations
      summary: Send a message in a conversation
      operationId: sendMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags:
        - Conversations
      summary: List conversation messages
      operationId: listMessages
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
        - name: before
          in: query
          required: false
          deprecated: true
          schema:
            type: string
            format: date-time
          description: Legacy timestamp filter for older messages. Prefer `cursor` for stable pagination.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque pagination cursor returned by the previous list response. Do not combine with `before`.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Messages ordered from oldest to newest within the returned page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/message-favorites:
    get:
      tags:
        - Conversations
      summary: List favorited messages
      operationId: listMessageFavorites
      description: Returns messages favorited by the current user. Messages deleted for the current user are not returned.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Favorited messages ordered by newest favorite first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageFavoriteListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: The request body or fields are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication failed or the bearer token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: A conflicting resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: The server could not complete the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    RecentDocsResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecentDoc"
...
    SyncResponse:
      type: object
      required:
        - serverTime
        - conversations
        - messages
        - nextCursor
        - hasMore
      properties:
        serverTime:
          type: string
          format: date-time
          description: Server-side snapshot time for the response. Legacy clients may reuse this as the next `since` value.
        conversations:
          type: array
          description: Current conversation summaries, including archived conversations and per-user settings.
          items:
            $ref: "#/components/schemas/ConversationSummary"
        messages:
          type: array
          description: Current page of messages created or updated after the supplied sync cursor, ordered oldest to newest by change order.
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: string
          description: Cursor to store after processing the page. When `hasMore` is true, call sync again immediately with this cursor.
        hasMore:
          type: boolean
          description: True when more changed messages remain in the current sync window.
...